backend/requirements.txt
# Web framework and related
flask==2.0.1
flask-cors==3.0.10
gunicorn==20.1.0

# HTTP and scraping
requests==2.26.0
beautifulsoup4==4.10.0
lxml==4.6.3

# Data processing
pandas==1.3.3
numpy==1.21.2

# Utilities
python-dotenv==0.19.0
apscheduler==3.8.0
psutil==5.8.0

# Time handling
pytz==2021.1